Store title formats for multiple sites.
Previously, as I focussed on the data structure of the tokens and their
interchange with the server, I neglected to remember that we have
different title formats for every given site.

This patch changes the state structure so that multiple sites' title
formats can be independently saved.

I have a question about the way I have done it since it feels like some
of the per-site stuff should be handled already for me instead of
requiring me to save an array or object with site ids.